split different screens into their own files
